ci: verify the crates against a 32-bit configured float width
The crates convert a different set of Rust floats under each configured
float width, and every archive the chain staged carried the 64-bit one,
so the 32-bit conversions were code nothing compiled. A verification
config builds mruby under MRB_USE_FLOAT32, and the suite and clippy run
against that archive too — the same shape the 64-bit integer width
already has.

It pins no integer width, so the leg differs from the validation build
in both axes at once. The first run earned its place immediately: twenty
test sites assumed `f64` boxes under every width, which is the claim
this width exists to refuse.

The configs' shared constants move to a file of their own, so a config
verifying a different ABI differs from the others in the ABI alone.

+++ crates/beni-tests/src/data_test.rs
@@ -387,7 +387,7 @@ fn marking_refuses_every_built_in_layout_so_its_instances_keep_it() {
         String::from_value(mrb.str_new(b"box").as_value()),
         Some("box".to_string())
     );
-    assert!(1.5f64.into_value(&mrb).is_float());
+    assert!(1.5f32.into_value(&mrb).is_float());
     let joined = mrb
         .load_string(b"BeniDataRefusedString.new('abc') + 'def'")
         .expect("the subclass still builds strings");
